requirements.txt
appnope==0.1.3
asgiref==3.5.2
asttokens==2.0.5
async-generator==1.10
attrs==21.4.0
backcall==0.2.0
certifi @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_884c889c-96af-444f-bd6d-daddb5e9a462ykj3l5n_/croots/recipe/certifi_1655968814730/work/certifi
cffi==1.15.1
charset-normalizer==2.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.4
decorator==5.1.1
Django==4.0.6
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-yasg==1.21.3
executing==0.8.3
h11==0.13.0
idna==3.3
inflection==0.5.1
ipython==8.4.0
itypes==1.2.0
jedi==0.18.1
Jinja2==3.1.2
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
outcome==1.2.0
packaging==21.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.30
ptyprocess==0.7.0
pure-eval==0.2.2
pybrowsers==0.5.0
pycparser==2.21
Pygments==2.12.0
PyJWT==2.4.0
PyMySQL==1.0.2
pyOpenSSL==22.0.0
pyparsing==3.0.9
PySocks==1.7.1
python-dotenv==0.20.0
pytz==2022.1
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
selenium==4.3.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
sqlparse==0.4.2
stack-data==0.3.0
traitlets==5.3.0
trio==0.21.0
trio-websocket==0.9.2
uritemplate==4.1.1
urllib3==1.26.10
wcwidth==0.2.5
webdriver-manager==3.8.0
wsproto==1.1.0
